sandlock-core: hoist openat2_in_root into shared sys::fs
