+++ crates/sandlock-core/src/sys/mod.rs
@@ -1,2 +1,3 @@
+pub mod fs;
 pub mod structs;
 pub mod syscall;
